% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_dup}
\alias{test_dup}
\title{Duplicates a given string}
\usage{
test_dup(x, i)
}
\arguments{
\item{x}{a character vector}

\item{i}{an integer vector}
}
\value{
a character vector
}
\description{
Duplicates a given string
}
\examples{
test_dup("A", 1)

}

