% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_paste}
\alias{test_paste}
\title{Concatenates all strings into a single one}
\usage{
test_paste(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
Concatenates all strings into a single one
}
\examples{
test_paste(letters)

}

