% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleRcppStringi-package.R
\docType{package}
\name{ExampleRcppStringi-package}
\alias{ExampleRcppStringi-package}
\alias{ExampleRcppStringi}
\title{A stringi-dependent Rcpp package example}
\description{
A stringi-dependent Rcpp package example
}
\author{
Marek Gagolewski
}
